Add support for multiple expectations that share the same 'probe'
It's now possible to feed a MarbleTest instance multiple expectation lines that share the same 'probe' - more specifically the same 'NothingElse' assertion. This is achieved by first testing all 'positive' expectation for a certain Moment, for all expectation lines, and only then test the 'negative' expectations - the Nothing Else for all lines.
Note that the negative expecation will 'fail' on the first line that has not 'ended' at the tested time, so the end user has control over the line that will appear in the exception message per moment in time.